(BLE) : preliminary read of values
(:) the value is only read once, and not always reliably. atm the value
is just printed in hex to the Log. After the value has been read, the
connection is closed. this should be done only when the device activity
is closed
(/=) the main activity now only show supported devices, but it can now
deal with multiple temperature sensors