(/=) Set parser delegate in ChatLogic
(+) Add callback in AsyncNetwork for when the message delivery fails
immediately (because there probably is no network available)
(!) Sorry about the confusion with the different interfaces.
ChatServerRawResponseInterface is the one implemented by the ChatLogic.
It has some lower level stuff like the clocks, but no message ids, as
they are only local.