(/=) Properly remove listeners
(/=) Do not fix the sync type once and for all.
(/=) Make some classes not public, as they are only internal to the
sub-package.
(/=) Fix logout issues: if the server kicks the user out, then any
subsequent message will be answered by a message with the „register“
command and „failure“ set to Not registered
(/=) Fix parsing of failure reasons